Use objectTemp instead of IRTemp
The simulator emits the temperature with a key of `objectTemp` instead of `IRTemp`. This commit fixes the Node RED flow to use the new key. It also turns on debug output for the temperature node only, so that the debug tab doesn't fill up as quickly as before.

Also add some useful debug information.